requirements.txt
microdata
beautifulsoup4
psycopg2
psycopg2-binary
SQLAlchemy
python-dotenv
